feat(ux): move Rate limits to /settings; drop Guard Connections tab (#1359) (#1410)
Rate limits are workspace-level default caps on proxy traffic, not
Guard-policy config — moved from Guard settings to workspace /settings
where they sit next to Vault / LLM Primitives / Appearance / Proxy /
Members. Guard's Connections tab was a bare link to /integrations —
removed since Guard's own top-level nav already covers that path.

- New apps/web/src/components/settings/RateLimitsPanel.tsx — extracted
  from RateLimitsCard + RateLimitPresets in the Guard settings page.
  guard.rateLimits.* API calls unchanged; only the surface moves.
- apps/web/src/app/(app)/settings/page.tsx: adds rate_limits tab
  (adminOnly) between Proxy and Members.
- apps/web/src/app/(app)/theguard/settings/page.tsx: -187 lines. Drops
  RateLimitsCard, RATE_LIMIT_PRESETS, RateLimitPresets, the rate_limits
  and connections tab entries and their panels.

Ref #1359. Follow-up to PR #1409.
